+++ src/media/visual/mod.rs
@@ -10,7 +10,6 @@
 pub mod color;
 #[cfg(feature = "draw")]
 pub mod draw;
-#[cfg(feature = "image")]
 pub mod image;
 // pub mod video;
 
@@ -18,7 +17,7 @@ crate::structural_mods! { // _pub_mods, _crate_internals
     _pub_mods {
         #[cfg(feature = "color")] pub use super::color::_all::*;
         #[cfg(feature = "draw")] pub use super::draw::_all::*;
-        #[cfg(feature = "image")] pub use super::image::_all::*;
+        pub use super::image::_all::*;
         // pub use super::video::_all::*;
     }
     _crate_internals {